consolidated components and implemented go fish popup
